+++ CHANGELOG.md
@@ -163,7 +163,7 @@ All notable changes to this project will be documented in this file. See [standa
 ### BREAKING CHANGES
 
 * autoCorrect, autoCapitalize are now removed.
-* `[ngOptionHighlight]` is now available from @ng-select/ng-option-highlight
+* `[ngOptionHighlight]` is now available from @ct-cue/ng-option-highlight
 * `[excludeGroupsFromDefaultSelection]` is now removed until better solution is found
 * `filterValue` is now available as `searchTerm`